+++ application/views/manage_organization.php
@@ -86,7 +86,7 @@
 									<a href="<?php echo base_url(); ?>master/add_organization/<?=$list->organization_id; ?>" class="btn btn-secondary btn-sm btn-icon icon-left">
 											Edit
 									</a>
-									<a href="<?php echo base_url(); ?>master/delete_organization/<?=$list->organization_id; ?>" class="btn btn-danger btn-sm btn-icon icon-left">
+									<a href="<?php echo base_url(); ?>master/delete_organization/<?=$list->organization_id; ?>" onClick="return confirm('Are you sure to delete this organization ? This will delete all the related records on this organization as well.')" class="btn btn-danger btn-sm btn-icon icon-left">
 											Delete
 									</a>
 								</td>
